pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "asset-bridge"
version = "0.1.0"
description = "Jewelry product image processing pipeline — cleanup, color variants, model shots, lifestyle settings"
requires-python = ">=3.11"
dependencies = [
    "typer>=0.12",
    "rich>=13",
    "rembg>=2.0",
    "onnxruntime>=1.17",
    "Pillow>=10",
    "pyyaml>=6",
    "python-dotenv>=1",
    "google-genai>=1",
    "openai>=1.40",
    "imagehash>=4",
    "scikit-image>=0.22",
    "numpy>=1.26",
    "streamlit>=1.38",
    "httpx>=0.27",
]

[project.optional-dependencies]
dev = ["ruff", "pytest", "pytest-asyncio"]
train = [
    "torch>=2.0",
    "torchvision>=0.15",
    "diffusers>=0.30",
    "transformers>=4.40",
    "accelerate>=0.30",
    "peft>=0.10",
    "safetensors>=0.4",
]

[project.scripts]
asset-bridge = "asset_bridge.cli:app"

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.hatch.build.targets.wheel]
packages = ["src/asset_bridge"]
